+++ CHANGELOG.md
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated `spicyweb\tinymce\assets\FieldAsset`; use `spicyweb\tinymce\assets\field\FieldAsset` instead
 - Deprecated `spicyweb\tinymce\assets\TinyMCEAsset`; use `spicyweb\tinymce\assets\tinymce\TinyMCEAsset` instead
 
+### Fixed
+- Fixed a bug where TinyMCE Fields could cause 404 errors when attempting to load translations
+
 ## 1.1.1 - 2022-10-26
 
 ### Fixed